gst: utils: Add a macro to use a GMutexLocker
The GMutexLocker uses GCC/CLANG C feature to implement a smart lock,
that unlocks the mutex when the GMutexLocker goes out of scope. This
could have been implemented in C++, but as this is already implemented
I decided to just reuse it. This is particularly handy to avoid gotos
in error handling cases and to prevent unbalanced locking.

Signed-off-by: Nicolas Dufresne <nicolas.dufresne@collabora.com>