gst: libcamerasrc: Implement minimal caps negotiation
This is not expected to work in every possible cases, but should be sufficient as
an initial implementation. What is does it that it turns the StreamFormats into
caps and query downstream caps with that as a filter.

The result is the subset of caps that can be used. We then keep the first
structure in that result and fixate using the default values found in
StreamConfiguration as a default in case a range is available.

We then validate this configuration and turn the potentially modified
configuration into caps that we push downstream. Note that we strust the order
in StreamFormats as being sorted best first, but this is not currently in
libcamera.

Signed-off-by: Nicolas Dufresne <nicolas.dufresne@collabora.com>